add list of tuples tests for Conform.Utils.merge/2
Prior to this commit we had no tests indicating the expected behaviour
of `Conform.Utils.merge/2` for lists of tuples, either generic or ips,
which meant we had no way to validate fix for GH-73 or guard against
future regression. This commit adds simple tests that something merged
with itself should still be itself, which fails for 2.0.0 and passes as
of 75be712.