Fix incorrect merging of lists
Conform merges configuration at various points, and does so as best it
can, by merging keyword lists/maps, and taking higher priority values
otherwise. A previous bugfix left things such that regular lists would
be merged in some cases, which is not the desired behaviour - for such
lists, the entire list on one side of the merge or the other should be
used.

See #117