adjust conform to Elixir 1.9
This commit provides fixing of deprecation warnings because of Elixir
version 1.9.

Following changes were made:

  * Mix.Releases.Logger was replaced with Conform.Logger as there is
no Mix.Releases.Logger anymore.
  * Mix.Dep.loaded/0 was replaced with Mix.Dep.load_on_environment/1
as the first one was deprecated and removed.
  * catch/rescue order fixed.

The biggest change is in the test for creating archive with dependencies.
Compliation of the fixtures projects was moved outside of
Mix.Project.in_project/4 because previous behaviour was changed in the
new Elixir and now it tries to build top-level (conform itself) dependencies
instead of dependencies of fixture apps from conform/test directory.

Mix.task.run is replaced with direct calling of `mix` tasks via System.cmd/2
to avoid these issues with dependencies and now everyting works as expected.

This shouldn't break anything, because for normal Elixir project it works
as expected as well.